requirements.txt
playwright
pytest
pytest-playwright
pytest-xdist
requests
pyyaml
streamlit
pandas
langchain
langchain-community
langchain-openai
langchain-core
python-dotenv
loguru
allure-pytest
fastapi
uvicorn
sqlalchemy
atlassian-python-api
